+++ src/module-freetype/Class_font.h
@@ -62,7 +62,7 @@ public:
 					const String &str, const Function *pFuncDeco);
 private:
 	FT_Error LoadAndDecorateChar(Environment &env, Signal &sig,
-				unsigned long codeUTF32, size_t idx, const Function *pFuncDeco);
+				UInt32 codeUTF32, size_t idx, const Function *pFuncDeco);
 	void DrawMonoOnImage(Image *pImage, int x, int y,
 				unsigned char *buffer, int width, int height, int pitch,
 				int xOffset, int yOffset);
